feat(services/users): make prisma work through context
